% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digipcr-package.R
\docType{package}
\name{digipcr-package}
\alias{digipcr-package}
\alias{digipcr}
\title{digital PCR processing tools}
\description{
Implementation of different methods to determine treshold
for classification and subsequent calculation of the
concentration in digital PCR samples.
}
